fix(automations): satisfy UI and type checks
